requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1

# LLM & AI Services
groq==0.4.1
pinecone-client==2.2.4

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
pydantic==2.5.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Async & Concurrency
asyncio-mqtt==0.16.1
websockets==12.0

# Utilities
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# Search & Tools
duckduckgo-search==3.9.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
